Move terrain definitions to jaxsim.terrain package
