+++ src/jaxsim/terrain/__init__.py
@@ -1,2 +1,2 @@
-from jaxsim.physics.algos import terrain
-from jaxsim.physics.algos.terrain import FlatTerrain, Terrain
+from . import terrain
+from .terrain import FlatTerrain, Terrain
