+++ src/jaxsim/terrain/terrain.py
